esbuild: Move duplicated config into a shared file
Put the shared esbuild logic in common/esbuild-config.mjs as a
runEsbuild() helper. Each extension now only lists its own entry
points and extra configs (qt-ui's webview) and passes in build and
context, since the shared file can't find 'esbuild' on its own from
the repo root. Also turn on --incremental tsc for all of them.
